revising conv code structure
