+++ src/owl/core/owl_core.h
@@ -10,6 +10,7 @@
 #include "owl_slicing.h"
 #include "owl_ndarray_contract.h"
 #include <stdio.h> // DEBUG
+#include <assert.h>
 
 
 /** Core function declaration **/
